+++ bin/index.js
@@ -1,22 +1,51 @@
 #!/usr/bin/env node
 
-const program = require('commander');
 const path = require('path');
-
 const parser = require('../');
 const generate = require('../lib/generate');
 
-program
-  .version(require('../package').version)
-  .option('-o, --out <file>', 'if instead of piping content you want it to be written to an html file locally please specify the relative path')
-  .parse(process.argv);
+let program = {};
+const args = process.argv.slice(2);
+
+args.forEach((arg, i) => {
+  switch (arg) {
+    case '-v':
+    case '--version':
+    case 'version':
+      console.log(`v${require('../package.json').version}`); // eslint-disable-line
+      process.exit(0);
+    break;
+    case '-h':
+    case '--help':
+    case 'help':
+      console.log(`` + // eslint-disable-line
+        `
+Usage: tap-html [options]
+
+Commands:
+  -h, --help, help                Output usage information
+  -v, --version, version          Output the version number
+
+Options:
+  -o, --outFile [path]            If instead of piping content you want it to be written to an html file locally please specify the relative path
+`);
+      process.exit(0);
+    break;
+    case '-o':
+    case '--out':
+      program['out'] = path.resolve(process.cwd(), args[i + 1]);
+    break;
+  }
+});
+
+const { out } = program;
 
 process.stdin
   .pipe(parser((res) => {
     // generate the html report
-    if(program.out) {
+    if(out) {
       // write the output to the specified location
-      generate(res, path.resolve(process.cwd(), program.out));
+      generate(res, out);
     } else {
       generate(res);
     }